updating data

+++ src/Component/MyToy.jsx
@@ -1,5 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from './AuthProvider';
+import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const MyToy = () => {
     const { user } = useContext(AuthContext)
@@ -11,27 +13,42 @@ const MyToy = () => {
     }, [user])
 
     const handelDelete = (id) => {
-        console.log(id)
-        fetch(`http://localhost:5000/delete/${id}`, {
-            method: "DELETE"
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/delete/${id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount === 1) {
+                            const remaining = myToy.filter(toy => toy._id !== id)
+                            setMyToy(remaining)
+                        }
+                    })
+            }
+
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount === 1) {
-                    const remaining = myToy.filter(toy => toy._id !== id)
-                    setMyToy(remaining)
-                }
-            })
+
+
     }
     return (
         <div>
-            <div className="overflow-x-auto">
+            <div className="overflow-x-auto ">
                 {
                     user ? <table className="table w-full">
-                       
+
                         <thead>
                             <tr>
                                 <th></th>
+                                <th>img</th>
                                 <th>Price</th>
                                 <th>Available quantity</th>
                                 <th>Detail description</th>
@@ -43,11 +60,12 @@ const MyToy = () => {
                             myToy.map((toy, i) => <tbody key={toy._id}>
                                 <tr>
                                     <th>{i + 1}</th>
+                                    <th><img className='w-[50px] rounded' src={toy.url} alt="" /></th>
                                     <td>${toy.price}</td>
-                                    <td>{toy.Available}</td>
+                                    <td>{toy.quantity}</td>
                                     <td>{toy.details}</td>
-                                    <td><button onClick={() => handelDelete(toy._id)}>Delete</button></td>
-                                    <td><button>Update</button></td>
+                                    <td><button className='bg-emerald-500 text-white p-1 rounded font-bold' onClick={() => handelDelete(toy._id)}>Delete</button></td>
+                                    <td><Link to={`/update/${toy._id}`}><button className='bg-sky-800 text-white p-1 rounded font-bold'>Update</button></Link></td>
                                 </tr>
                             </tbody>)
                         }
@@ -55,6 +73,7 @@ const MyToy = () => {
                     </table> : ''
                 }
             </div>
+
         </div>
     );
 };
